second commit

+++ routes/auth.js
@@ -1,31 +1,37 @@
 const express = require("express");
 const User = require("../models/User");
 const bcrypt = require("bcrypt");
+const BASE_URL = require('../routes/__config')
 const { sign } = require("jsonwebtoken");
 const {
   validateUser,
   validateUserUpdate,
   validateToken,
+  upload_files,
+  
 } = require("./validateUser");
+const {userSerializer} = require('./__functions')
 const router = express.Router();
 
 // register
 router.post("/register", async (req, res) => {
   // validating the fields
-  validateUser(req.body);
+  // console.log(req.body);
+  let error = validateUser(req.body)
+  if(error.length>0)return res.send(error)
 
   const { name, email, phone, password, usertype } = req.body;
 
   // check for duplicate db registry
   const check = await User.findOne({ email, usertype });
-  if (check) return res.status(409).send({ error: "user already exist" });
-
+  if (check) return res.status(400).send({ error: "user already exist" });
+ 
   // go on save user
   const salt = await bcrypt.genSalt(10);
   const hashpassword = await bcrypt.hash(password, salt);
   const user = new User({
     name,
-    email,
+    email, 
     phone,
     usertype,
     password: hashpassword,
@@ -36,15 +42,19 @@ router.post("/register", async (req, res) => {
   try {
 
     await user.save();
-    res.send('created').status(201);
+    res.send(user).status(200);
   } catch (error) {
-    res.status(400).send(error);
+    res.status(400).send([error.messege]);
     console.log(error);
   }
 });
+
+
+
 // ............................................................................................................................................................................
 // login
 router.post("/login", async (req, res) => {
+  
   const { email, password, usertype } = req.body;
   try {
     const user = await User.findOne({ email, usertype });
@@ -69,19 +79,10 @@ router.get("/me", validateToken, async (req, res) => {
   const user = req.user;
   try {
     const exist = await User.findById(user.id);
+    // return res.send(exist)
     if (!exist) return res.send({ error: "no such user user" });
 
-    let doc = {
-      _id: exist._id,
-      name: exist.name,
-      email: exist.email,
-      phone: exist.phone,
-      active: exist.active,
-      avater: exist.avater,
-      usertype: exist.usertype,
-      createdAt: exist.createdAt,
-      updatedAt: exist.updatedAt,
-    }
+    let doc = userSerializer(exist)
     res.send(doc);
   } catch (error) {
     console.log(error);
@@ -96,17 +97,7 @@ router.get("/user/:id", validateToken, async (req, res) => {
     const user = await User.findById(req.params.id);
     if (!user) return res.send({ error: "no such user exist" });
 
-   let doc = {
-      _id: user._id,
-      name: user.name,
-      email: user.email,
-      phone: user.phone,
-      active: user.active,
-      avater: user.avater,
-      usertype: user.usertype,
-      createdAt: user.createdAt,
-      updatedAt: user.updatedAt,
-    }
+   let doc = userSerializer(user)
 
     res.send(doc);
   } catch (error) {
@@ -116,27 +107,18 @@ router.get("/user/:id", validateToken, async (req, res) => {
 // ............................................................................................................................................................................
 
 // get all the user
-router.get("/users", validateToken, async (req, res) => {
+router.get("/", validateToken, async (req, res) => {
   try {
-    const users = await User.find();
+    const users = await User.find().sort({usertype:-1}).sort({avater:-1});
 
 
     if (!users) return res.status(400).send({ error: "error finding users" });
     let user_list =[]
     for(let doc of users){
-      let obj ={}
-      obj._id = doc._id,
-      obj.name = doc.name,
-      obj.email = doc.email,
-      obj.phone = doc.phone,
-      obj.active = doc.active,
-      obj.avater = doc.avater,
-      obj.usertype = doc.usertype,
-      obj.createdAt = doc.createdAt,
-      obj.updatedAt = doc.updatedAt,
+    
 
       // puseh to empty arrau
-      user_list.push(obj)
+      user_list.push(userSerializer(doc))
     }
 
     return res.send(user_list);
@@ -147,42 +129,74 @@ router.get("/users", validateToken, async (req, res) => {
 
 // ............................................................................................................................................................................
 // delete auser
-router.delete("/delete/:id", validateToken, async (req, res) => {
-  try {
-    const deleteUser = await User.deleteOne({ _id: req.params.id });
-    if (!deleteUser) return res.status(400).send("failed");
-    res.send("success");
-  } catch (error) {
-    console.log(error);
+router.delete("/:id", validateToken, async (req, res) => {
+  const {usertype} = req.user
+  if(usertype==='super admin'){
+
+    try {
+      
+      const deleteUser = await User.deleteOne({ _id: req.params.id });
+      if (!deleteUser) return res.status(400).send("failed");
+      res.send("deleted");
+    } catch (error) {
+      console.log(error);
+    }
+  }else{
+    res.send('Unauthorised action')
   }
 });
 // ............................................................................................................................................................................
 
 // upate auser
-router.put("/update", validateToken, async (req, res) => {
-  const { id, name, email, phone, usertype, active } = req.body;
-  try {
-    // validating the fields
+router.patch("/", validateToken, async (req, res) => {
+  const {id, usertype} = req.user
+  if(usertype ==='admin'){
+
+    
+    const { name, email, phone} = req.body;
+    try {
+      // validating the fields
     validateUserUpdate(req.body);
 
     // find one and update
-      let doc = await User.findOneAndUpdate({_id:id}, {name, email, phone, usertype, active}, {new:true});
-      doc = {
-        _id: doc._id,
-        name: doc.name,
-        email: doc.email,
-        phone: doc.phone,
-        active: doc.active,
-        avater: doc.avater,
-        usertype: doc.usertype,
-        createdAt: doc.createdAt,
-        updatedAt: doc.updatedAt,
-      }
-      // delete doc.password
-      return res.send(doc)
+    let doc = await User.findOneAndUpdate({_id:id}, {name, email, phone}, {new:true});
+    doc = {
+      name: doc.name,
+      email: doc.email,
+      phone: doc.phone,
+      
+    }
+    // delete doc.password
+    return res.send(doc)
   } catch (error) {
     console.log(error);
   }
+}else{
+  res.send('400 bad request').status(400)
+}
+});
+// upate auser
+router.patch("/admin/:id", validateToken, async (req, res) => {
+ 
+    const { name, email, phone} = req.body;
+    try {
+      // validating the fields
+    validateUserUpdate(req.body);
+
+    // find one and update
+    let doc = await User.findOneAndUpdate({_id:id}, {name, email, phone}, {new:true});
+    doc = {
+      name: doc.name,
+      email: doc.email,
+      phone: doc.phone,
+      
+    }
+    // delete doc.password
+    return res.send(doc)
+  } catch (error) {
+    console.log(error);
+  }
+
 });
 // ............................................................................................................................................................................
 
@@ -190,28 +204,46 @@ router.put("/update", validateToken, async (req, res) => {
 
 // update user to be inative
 
-// router.patch('/active/:id', validateToken, async (req, res)=>{
-//   const user = req.user
-//   const id = req.params.id
-//   return res.send({user,id})
+router.patch('/active/:id', validateToken, async (req, res)=>{
+  const user = req.user
+  const id = req.params.id
+  // return res.send({user,id})
 
-//   if(user.usertype === 'admin'){
-//     const check = await User.findById(id)
+  if(user.usertype === 'super admin'){
+    const check = await User.findById(id)
 
-//     // chech the current active status
-//     const active = check.active == true?false:true 
+    // chech the current active status
+    const active = check.active == 'active'?'inactive':'active' 
+    
+    // go ahead updating user
+    await User.findOneAndUpdate({_id:id},{$set:{active:active}}, {new:true});
+    return res.send('Done') 
+    
 
-//     // go ahead updating user
-//     let doc = await User.findOneAndUpdate({_id:id},{usertype:active}, {new:true});
-//     return res.send(doc).status(201)
 
-
-//   }
-//   res.send({error:'you cannot perform this action'})
+  }
+  res.send('you cannot perform this action')
 
   
 
 
-// })
+})
+
+router.post('/avater', validateToken, async (req,res)=>{
+  const {id} = req.user
+  const file = req.files
+  if(req.files){
+    const avater = await upload_files(file.avater)
+
+    // saving it to db
+    const _avater = await User.updateOne({_id:id}, {$set:{avater}})
+
+    console.log(_avater);
+    res.send('sent').status(200);
+
+  }
+    
+})
+
 
 module.exports = router;
